Added alerts and space
I added an alert for the sinus_count and covid_count because I realized that was part of my code in app 3, but it was missing in this one, so I am hoping it will help the percentage to work. I also added a line of space after the resultText.innerHTML code and the bracket.